requirements.txt
arabic-reshaper==3.0.0
chardet==5.2.0
pillow==10.2.0
python-bidi==0.4.2
reportlab==4.1.0
six==1.16.0
